requirements.txt
Flask
Flask-Cors
googletrans==4.0.0-rc1
python-dotenv
yt-dlp
google-api-python-client
isodate
google-generativeai
gunicorn
Flask-SQLAlchemy
Flask-Migrate
PyJWT
Authlib
pydub